+++ airline-prompts/src/main/java/com/github/rvesse/airline/prompts/formatters/QuestionFormat.java
@@ -13,7 +13,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
 package com.github.rvesse.airline.prompts.formatters;
 
 import org.apache.commons.collections4.CollectionUtils;
@@ -23,15 +22,15 @@ import com.github.rvesse.airline.io.printers.UsagePrinter;
 import com.github.rvesse.airline.prompts.Prompt;
 import com.github.rvesse.airline.prompts.builders.ListFormatBuilder;
 
+import java.io.PrintWriter;
+import java.util.Arrays;
+import java.util.stream.Collectors;
+
 /**
  * Prompt format for simple questions with either a free-form response or with a limited number of options
  *
- * @param <TOption>
- *            Option type
  */
-public class QuestionFormat<TOption> implements PromptFormatter {
-
-    private final int columns;
+public class QuestionFormat extends AbstractPromptFormat {
 
     /**
      * Creates a new question format with default columns
@@ -42,23 +41,37 @@ public class QuestionFormat<TOption> implements PromptFormatter {
 
     /**
      * Creates a new question format with the specified columns
-     * 
-     * @param columns
-     *            Columns
+     *
+     * @param columns Columns
      */
     public QuestionFormat(int columns) {
-        this.columns = columns;
+        super(columns);
     }
 
     @Override
     public <T> void displayPrompt(Prompt<T> prompt) {
-        UsagePrinter printer = new UsagePrinter(prompt.getProvider().getPromptWriter(), this.columns);
+        PrintWriter writer = prompt.getProvider().getPromptWriter();
+        UsagePrinter printer = new UsagePrinter(writer, this.columns);
         if (CollectionUtils.isNotEmpty(prompt.getOptions())) {
-            printer.append(String.format("%s? [%s] ", prompt.getMessage(), StringUtils.join(prompt.getOptions(), "/")));
+            printer.append(
+                    String.format("%s? [%s]", prompt.getMessage(), formatOptions(prompt)));
         } else {
-            printer.append(String.format("%s? ", prompt.getMessage()));
+            printer.append(String.format("%s?", prompt.getMessage()));
         }
         printer.flush();
+        // NB - UsagePrinter strips trailing whitespace so have to explicitly add the whitespace after the question
+        //      prompt otherwise the prompt looks a bit clunky in the console
+        writer.print(' ');
+        writer.flush();
+    }
+
+    private <T> String formatOptions(Prompt<T> prompt) {
+        String optionsString = prompt.getOptions().stream().map(this::formatOption).collect(Collectors.joining("/"));
+        if (prompt.getDefaultOption() == null) {
+            return optionsString;
+        } else {
+            return optionsString + " [" + prompt.getDefaultOption() + "]";
+        }
     }
 
 }